monitor/requirements.txt
speedtest-cli==2.1.3
flask==3.0.3
sqlalchemy==2.0.36